src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Northwind Commerce knowledge assistant"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.0.14",
    "openai>=1.0.0",
    "sqlite-vec>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "PyJWT>=2.8.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "logfire[fastapi]>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "ruff>=0.1.0",
    "ty>=0.0.17",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.setuptools]
packages = [
    "domain",
    "domain.infrastructure",
    "application",
    "application.use_cases",
    "application.infrastructure",
    "presentation",
    "presentation.routes",
    "presentation.infrastructure",
]
py-modules = ["main", "config", "logging_config"]

[tool.ty.environment]
root = ["."]

[tool.ty.rules]
unresolved-import = "ignore"
unresolved-attribute = "ignore"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
